==> SINGLE-CELL SIMULATION/2 - Run differential methods - batch/0 - prepare_brie.R <==
alias R='R_LIBS=/home/stiber/R/x86_64-pc-linux-gnu-library/4.3 /home/Shared/simone/DifferentialRegulation/software/R-4.3.0/bin/R'

cd /home/Shared_sherborne/simone/Diff_Velo/NEW_simulation/06_alevin_fry_USA

R

# clean environment
rm(list = ls())

# load packages
suppressPackageStartupMessages({
  # load libraries
  library(anndata)
  library(fishpond)
  library(SingleCellExperiment)
})

# load abundances
sample_ids <- 1:4
cell_types <- c("Adipocytes", "Epithelial_cells", "Hepatocytes")

load_abundances <- function (sample_id, cell_type) {
  path <- paste0("simulation_batch/", paste0("normal", sample_id),
                 "_", cell_type)
  temp <- loadFry(path, outputFormat = "raw", nonzero = FALSE, quiet = TRUE)
  temp$sample_id <- sample_id
  colnames(temp) <- paste0(sample_id, ".", colnames(temp))
  return (temp)
}

clusters = c("Adipocytes", "Epithelial_cells", "Hepatocytes")

for(cl in clusters){
  sces <- lapply(sample_ids, load_abundances, cell_type = cl)
  
  sces <- unlist(sces)
  sce <- do.call("cbind", sces)
  assays(sce)
  
  assays(sce)$TOT_counts <- assays(sce)$spliced + assays(sce)$unspliced + assays(sce)$ambiguous
  
  # remove undetected genes
  sce <- sce[rowSums(assays(sce)$TOT_counts > 0) > 0, ]
  
  # set group parameter
  GROUP <- c("A", "B", "A", "B")
  batch = c("1", "1", "2", "2")
  
  # FILTER SCE BEFORE RUNNING BRIE2:
  min_count = 10
  
  # set group attribute
  sce$group <- ifelse(sce$sample_id %in% which(GROUP == "A"), "A", "B")
  sce$batch <- ifelse(sce$sample_id %in% which(batch == "2"), "2", "1")
  
  # remove lowly expressed genes: at least 10 non-zero cells:
  filter <- rowSums(assays(sce)$TOT_counts[, sce$group == "A"]) >= min_count & 
    rowSums(assays(sce)$TOT_counts[, sce$group == "B"]) >= min_count
  
  sce <- sce[filter, ]
  
  assays(sce)$TOT_counts = NULL
  assays(sce)$counts = assays(sce)$spliced
  
  library(zellkonverter)
  ad = SCE2AnnData(sce, X_name = "counts")
  ad$write_h5ad(paste0("BRIE2_USA/simulation_batch/cell_info_",
                       cl, ".h5ad"))
  
  DF = data.frame(index = colnames(sce), 
                  is_A = ifelse(sce$group == "A", 1, 0),
                  is_batch = ifelse(sce$batch == "2", 1, 0))

  write.table(DF,
              file = paste0("BRIE2_USA/simulation_batch/cell_info_", 
                            cl, ".tsv"),
              quote = FALSE,
              sep = "\t", row.names = FALSE)
}
